Move text chat to its own components folder
